% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grumpy-package.R
\docType{package}
\name{grumpy-package}
\alias{grumpy}
\alias{grumpy-package}
\title{grumpy: Read 'NumPy' '.npy' and '.npz' Files in R}
\description{
Lightweight way to read 'NumPy' '.npy' and '.npz' files in R. All data types supported by 'NumPy', with all sizes (converted internally to R native size), both C and Fortran order, and any shape, up to an arbitrary number of dimensions, are supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hugogruson.fr/grumpy/}
  \item \url{https://github.com/Bisaloo/grumpy}
  \item Report bugs at \url{https://github.com/Bisaloo/grumpy/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [copyright holder]
  \item Mike Smith (Original author of portions of the C code migrated from the Rarr package) [copyright holder]
}

Other contributors:
\itemize{
  \item German Network for Bioinformatics Infrastructure - de.NBI [funder]
}

}
\keyword{internal}
