% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npy.R
\name{parse_npy_datatype}
\alias{parse_npy_datatype}
\title{Parse a NumPy Array-protocol type strings}
\usage{
parse_npy_datatype(descr)
}
\arguments{
\item{descr}{A NumPy dtype description string, or a list of such strings fo
structured dtypes}
}
\value{
A list containing the parsed data type information, including the base
type, the number of bytes, and the endianness
}
\description{
Parse a NumPy Array-protocol type strings
}
\examples{
parse_npy_datatype(">i8")
parse_npy_datatype("|b1")
parse_npy_datatype(list(c("r", "<i8"), c("g", "<i8"), c("b", "<i8")))

}
