% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npy.R
\name{read_npy}
\alias{read_npy}
\title{Read a .npy file}
\usage{
read_npy(file)
}
\arguments{
\item{file}{Path to the .npy file}
}
\value{
An array containing the data from the .npy file
}
\description{
Read a .npy file
}
\examples{
read_npy(
  system.file("extdata", "test.npy", package = "grumpy")
)
}
