% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npz.R
\name{read_npz}
\alias{read_npz}
\title{Read a .npz file}
\usage{
read_npz(file)
}
\arguments{
\item{file}{Path to the .npz file}
}
\value{
A list of arrays containing the data from the .npz file
}
\description{
Read a .npz file
}
\examples{
read_npz(
  system.file("extdata", "test.npz", package = "grumpy")
)
}
